# core-lending-collateral-management-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-collateral-management
  description: Collateral & Guarantee Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/general-ledger
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/guarantees/{guaranteeRecordId}/valuations/{guaranteeValuationId}:
    get:
      tags:
        - GuaranteeValuation
      summary: Get a Guarantee Valuation by ID
      operationId: getById_1
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: guaranteeValuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GuaranteeValuationDTO'
    put:
      tags:
        - GuaranteeValuation
      summary: Update a Guarantee Valuation
      operationId: update_1
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: guaranteeValuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuaranteeValuationDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/GuaranteeValuationDTO'
    delete:
      tags:
        - GuaranteeValuation
      summary: Delete a Guarantee Valuation
      operationId: delete_1
      parameters:
        - name: guaranteeRecordId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: guaranteeValuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/collateral-cases/{collateralCaseId}/assets/{collateralAssetId}/valuations/{collateralValuationId}:
    get:
      tags:
        - CollateralValuation
      summary: Get a Collateral Valuation by ID
      operationId: getById_8
      parameters:
        - name: collateralCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralValuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollateralValuationDTO'
    put:
      tags:
        - CollateralValuation
      summary: Update a Collateral Valuation
      operationId: update_8
      parameters:
        - name: collateralCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralValuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollateralValuationDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollateralValuationDTO'
    delete:
      tags:
        - CollateralValuation
      summary: Delete a Collateral Valuation
      operationId: delete_8
      parameters:
        - name: collateralCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralValuationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/collateral-cases/{collateralCaseId}/assets/{collateralAssetId}/parties/{collateralPartyId}:
    get:
      tags:
        - CollateralParty
      summary: Get a Collateral Party by ID
      operationId: getById_9
      parameters:
        - name: collateralCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralPartyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollateralPartyDTO'
    put:
      tags:
        - CollateralParty
      summary: Update a Collateral Party
      operationId: update_9
      parameters:
        - name: collateralCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralPartyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollateralPartyDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollateralPartyDTO'
    delete:
      tags:
        - CollateralParty
      summary: Delete a Collateral Party record
      operationId: delete_9
      parameters:
        - name: collateralCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralPartyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/collateral-cases/{collateralCaseId}/assets/{collateralAssetId}/valuations:
    get:
      tags:
        - CollateralValuation
      summary: List or Search Valuations for an Asset
      operationId: findAll_8
      parameters:
        - name: collateralCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: filterRequest
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestCollateralValuationDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseCollateralValuationDTO'
    post:
      tags:
        - CollateralValuation
      summary: Create a Valuation under an Asset
      operationId: create_8
      parameters:
        - name: collateralCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollateralValuationDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollateralValuationDTO'
  /api/v1/collateral-cases/{collateralCaseId}/assets/{collateralAssetId}/parties:
    get:
      tags:
        - CollateralParty
      summary: List or Search Collateral Parties
      operationId: findAll_9
      parameters:
        - name: collateralCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: filterRequest
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestCollateralPartyDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseCollateralPartyDTO'
    post:
      tags:
        - CollateralParty
      summary: Create a Collateral Party record
      operationId: create_9
      parameters:
        - name: collateralCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: collateralAssetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollateralPartyDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollateralPartyDTO'
...
components:
  schemas:
...
    GuaranteeValuationDTO:
      required:
        - assessedRiskAmount
        - guaranteeRecordId
        - rationale
        - riskGrade
      type: object
      properties:
        guaranteeValuationId:
          type: string
          format: uuid
          readOnly: true
        guaranteeRecordId:
          type: string
          format: uuid
        assessedRiskAmount:
          minimum: 0
          exclusiveMinimum: false
          type: number
        riskGrade:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
        rationale:
          maxLength: 2000
          minLength: 1
          type: string
        valuationDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CollateralValuationDTO:
      required:
        - appraisedValue
        - collateralAssetId
        - currencyCode
        - valuationDate
        - valuationMethod
        - valuationProvider
      type: object
      properties:
        collateralValuationId:
          type: string
          format: uuid
          readOnly: true
        collateralAssetId:
          type: string
          format: uuid
        appraisedValue:
          minimum: 0
          exclusiveMinimum: true
          type: number
        valuationMethod:
          type: string
          enum:
            - MARKET_COMPARISON
            - INCOME_APPROACH
            - COST_APPROACH
            - OTHER
        valuationProvider:
          maxLength: 200
          minLength: 1
          type: string
        appraiserName:
          maxLength: 255
          minLength: 0
          type: string
        appraiserLicense:
          maxLength: 100
          minLength: 0
          type: string
        valuationDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        currencyCode:
          type: string
          enum:
            - EUR
            - USD
            - GBP
        marketValue:
          minimum: 0
          exclusiveMinimum: false
          type: number
        rentalValue:
          minimum: 0
          exclusiveMinimum: false
          type: number
        replacementCost:
          minimum: 0
          exclusiveMinimum: false
          type: number
        landValue:
          minimum: 0
          exclusiveMinimum: false
          type: number
        buildingValue:
          minimum: 0
          exclusiveMinimum: false
          type: number
        salvageValue:
          minimum: 0
          exclusiveMinimum: false
          type: number
        comparableAssets:
          type: string
        methodology:
          maxLength: 2000
          minLength: 0
          type: string
        assumptions:
          maxLength: 2000
          minLength: 0
          type: string
        limitations:
          maxLength: 2000
          minLength: 0
          type: string
        requiresRepairs:
          type: boolean
        requiredRepairs:
          maxLength: 2000
          minLength: 0
          type: string
        repairCost:
          minimum: 0
          exclusiveMinimum: false
          type: number
        notes:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CollateralPartyDTO:
      required:
        - collateralAssetId
        - partyId
        - roleCode
      type: object
      properties:
        collateralPartyId:
          type: string
          format: uuid
          readOnly: true
        collateralAssetId:
          type: string
          format: uuid
        partyId:
          type: string
          format: uuid
        roleCode:
          type: string
          enum:
            - OWNER
            - CO_OWNER
            - OCCUPANT
            - CUSTODIAN
        ownershipPercentage:
          maximum: 100
          exclusiveMaximum: false
          minimum: 0
          exclusiveMinimum: false
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
...
    FilterRequestCollateralValuationDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/CollateralValuationDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationResponseCollateralValuationDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/CollateralValuationDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
    FilterRequestCollateralPartyDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/CollateralPartyDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationResponseCollateralPartyDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/CollateralPartyDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...